add folders and files
