--- requirements.txt
@@ -1,13 +1,0 @@
-jupyter==1.0.0
-numpy==1.26.4
-pandas==1.4.3
-requests==2.28.1
-SQLAlchemy==1.4.39
-pyarrow==8.0.0
-pg8000==1.29.1
-pytest==7.1.2
-pylint==2.14.4
-pyyaml==6.0
-Jinja2==3.1.2
-schedule==1.1.0
-python-dotenv==1.0.0
