--- stock_bars.yaml
@@ -1,4 +1,0 @@
-name: alpaca
-config: 
-  stock_symbol_relative_path: etl_project/data/top_tech_stock_symbol.csv
-  load_method: upsert
