--- top_tech_stock_symbol.csv
@@ -1,11 +1,0 @@
-﻿Company,Symbol,Exchange
-Apple,AAPL,NASDAQ
-Microsoft,MSFT,NASDAQ
-Amazon,AMZN,NASDAQ
-Alphabet (Google),GOOGL,NASDAQ
-Meta Platforms (Facebook),META,NASDAQ
-Tesla,TSLA,NASDAQ
-NVIDIA,NVDA,NASDAQ
-Adobe,ADBE,NASDAQ
-Salesforce,CRM,NYSE
-Intel,INTC,NASDAQ
